requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==4.2.17
django-filter==24.3
django-jazzmin==3.0.1
django-semantic-admin==0.5.0
django-semantic-forms==0.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
